Move component MonitoringCommentCard from 'components' to 'components/index/SiteTopUpper'
